Remove more dead MP code: sanitize_data ConcurrentMap case, mptypes pruning
- Remove Agent-specific ConcurrentMap case from sanitize_data.go
  (tdns-mp uses its own Agent type; this path was never hit)
- Remove dead MP types and fields from structs.go
- Remove dead registerPeerAgents from main_initfuncs.go
- Remove unused Combiner field from config.go
